fix(HTTP Request Node): Cleanup circular references in response (#6590)
