+++ packages/workflow/src/index.ts
@@ -31,6 +31,7 @@ export {
 	sleep,
 	fileTypeFromMimeType,
 	assert,
+	removeCircularRefs,
 } from './utils';
 export {
 	isINodeProperties,
